Add files via upload
So far I am able to produce a sequence with different angled grates, when I input the degrees directly/insert a float. 

at the bottom is a non connected timer/random expression to these degrees (0,45,90,135), but when I input this to the angle of the grate it doesn't work, it changes the angle sequence. I tried making the timer to produce random trigger at longer periods to ensure the timer is not retriggering a new random each time, but also doesn't help. I think the random sequence and angles are being produced correctly, checked the output at least...