Add files via upload
circle location moves, grate moves. Start stim 0 and 1 (for grate) are somehow being produced even when min and max stim ID are 5 and 8. Start stim 8 is being produced when min and max stim ID are 0 to 3. Racing subjects? need to go over how grouping icons works to mitigate this issue. Tried to group items in the sequence, but this created problems also. Wanted to combine end of sequence at some point, as the Trial Num is counting separately the circle and grate sequences.