Add files via upload
circle location moves, grate moves. Start stim 0 and 1 (for grate) are somehow being produced even when min and max stim ID are 5 and 8. Start Stim 8 is now acting as expected. Potentially still a racing subjects issue, although I adjusted for this and don’t really see where now? need to go over how grouping icons works to mitigate this issue. Tried to group items in the sequence, but this created problems also. Wanted to combine end of sequence at some point, as the Trial Num is counting separately the circle and grate sequences.