Phase3_1IncorrectGrate
adjusted so an incorrect grate replaces the last element of a 3-trial Stimuli.

where is set with WhichPositionStimuli, every time the TrialNum updates. (for now 2/last element). want to adjust so this can be randomized, and need notequal to work in the StimID logic. NotEqual doesn't allow input of a value right now, only true false?