arduino serial comm
arduino code for bonsai communication in response window. detects cap sensor info (pin 2) and produce a tone or white noise (pin 11) based on response or no response in the window.  need to fix so that the next trial starts after the tone or white noise/timeout. right now it overlaps.